Update to 1.2.13

+++ mewwalletbl/src/main/java/com/myetherwallet/mewwalletbl/data/database/TotalBalance.kt
@@ -1,7 +1,9 @@
 package com.myetherwallet.mewwalletbl.data.database
 
+import java.math.BigDecimal
+
 data class TotalBalance(
     val amount: Long,
     val decimals: Int,
-    val price: Double
+    val price: BigDecimal
 )
